requirements.txt
langgraph==0.2.28
langchain==0.3.7
langchain-openai==0.2.8
openai==1.54.0
llama-index==0.11.22
chromadb==0.5.18
sentence-transformers==3.2.1
pypdf==5.1.0
mem0ai==0.1.29
fastapi==0.115.4
uvicorn==0.32.0
python-hcl2==4.3.4
checkov==3.2.258
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
PyGithub==2.4.0
rich==13.9.4
streamlit==1.40.1